# api/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateEmergencyContactInput:
    properties:
      is_priority:
        type: boolean
      name:
        type: string
      phone:
        type: string
      relation:
        enum:
        - family
        - friend
        - colleague
        - neighbor
        - other
        type: string
    required:
    - name
    - phone
    - relation
    type: object
...
  dto.EmergencyAlertInput:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      message:
        type: string
    required:
    - latitude
    - longitude
    type: object
  dto.EmergencyAlertResponse:
    properties:
      contacts_notified:
        type: integer
      message:
        type: string
      notified_contacts:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
  dto.EmergencyContactResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_priority:
        type: boolean
      name:
        type: string
      phone:
        type: string
      relation:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdateEmergencyContactInput:
    properties:
      is_priority:
        type: boolean
      name:
        type: string
      phone:
        type: string
      relation:
        enum:
        - family
        - friend
        - colleague
        - neighbor
        - other
        type: string
    required:
    - name
    - phone
    - relation
    type: object
...
  util.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: integer
          message: {}
        type: object
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@riskplace.ao
    name: API Support
    url: http://www.riskplace.ao
  description: |-
    This is the API documentation for the Risk Place Angola application.

    ## Environments
    - **Development**: https://risk-place-angola-904a.onrender.com
    - **Local**: http://localhost:8000

    ## WebSocket Integration Guide
    For detailed instructions on integrating WebSocket in mobile applications, please refer to our [WebSocket Notification Guide](https://github.com/risk-place-angola/backend-risk-place/blob/develop/docs/MOBILE_WEBSOCKET_INTEGRATION.md)
  title: Risk Place Angola API
  version: 1.0.0
paths:
...
  /emergency/alert:
    post:
      consumes:
      - application/json
      description: Send an emergency SMS alert with location to all priority emergency
        contacts
      parameters:
      - description: Emergency alert data with location
        in: body
        name: alert
        required: true
        schema:
          $ref: '#/definitions/dto.EmergencyAlertInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmergencyAlertResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send emergency alert to all priority contacts
      tags:
      - emergency
...
  /users/me/emergency-contacts:
    get:
      description: Retrieve all emergency contacts configured by the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EmergencyContactResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all emergency contacts for the current user
      tags:
      - emergency-contacts
    post:
      consumes:
      - application/json
      description: Add a new emergency contact for the authenticated user
      parameters:
      - description: Emergency contact data
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEmergencyContactInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EmergencyContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new emergency contact
      tags:
      - emergency-contacts
  /users/me/emergency-contacts/{id}:
    delete:
      description: Delete an emergency contact for the authenticated user
      parameters:
      - description: Emergency Contact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an emergency contact
      tags:
      - emergency-contacts
    put:
      consumes:
      - application/json
      description: Update an existing emergency contact for the authenticated user
      parameters:
      - description: Emergency Contact ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated emergency contact data
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEmergencyContactInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmergencyContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an emergency contact
      tags:
      - emergency-contacts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  DeviceID:
    description: Unique device identifier for tracking and analytics.
    in: header
    name: X-Device-ID
    type: apiKey
swagger: "2.0"